removed pyttsx3 in favor of piperTTS
the main remaining difference between running on Windows and Linux is an audio issue where Windows doesn't work well with simpleaudio installed and Linux (or just Raspberry Pi) has trouble with pyaudio dependencies.  So I made a note for Linux users to try installing simpleaudio, warning that Windows users should not try this.