add more voice assistant feedback
The voice assistant now provides an introduction message when ready to start listening.  It also prints what was heard when it detects speech that doesn't contain a wakeword.  Additionally, the user can turn off the assistant using the new exit commands listed in intent.  Finally, I removed "hey pi" and "hey pie" as default wakewords as they are unreliable on both of my devices.